save team color in mainmanager script
create a serializable class in main manager script that has a color type. add SaveColor function in main manager class that creates an instance of a save data and sets that team color to the main managers team color then puts it to jason data saved in the presistent path. the load color just does this backwards after checking to see if the file exist